requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.9.0
drf-spectacular==0.27.2
CherryPy>=18.0.0
pexpect>=4.9.0
